+++ bench/type4/adversarial/cases/swift_flat_map_aggregate_custom_filter/overload.swift
@@ -1,0 +1,3 @@
+extension Array where Element == Int {
+    func filter(_ predicate: (Int) -> Bool) -> [Int] { [] }
+}